Fix bugs from group chatting

+++ src/main/java/com/autohub/serviceImpl/VehicleServiceImpl.java
@@ -11,9 +11,6 @@ import com.autohub.service.VehicleService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -381,10 +378,18 @@ public class VehicleServiceImpl implements VehicleService {
 
 
     @Override
-    public VehicleResponseDTO updateVehicle(Long id, VehicleRequestDTO request) {
+    public VehicleResponseDTO updateVehicle(Long id, VehicleRequestDTO request,Long loggedInDealerId) throws AccessDeniedException {
 
         Vehicle vehicle = vehicleRepository.findById(id).orElseThrow(() ->
                         new ResourceNotFoundException("Vehicle not found with id: " + id));
+
+
+        if (!vehicle.getDealer().getId()
+                .equals(loggedInDealerId)) {
+
+            throw new AccessDeniedException(
+                    "You are not authorized to update this vehicle");
+        }
 
         vehicle.setBrand(request.getBrand());
         vehicle.setModel(request.getModel());
@@ -424,11 +429,18 @@ public class VehicleServiceImpl implements VehicleService {
 
 
     @Override
-    public VehicleResponseDTO updateVehicleStatus(Long id,VehicleStatusRequestDTO request) {
+    public VehicleResponseDTO updateVehicleStatus(Long id,VehicleStatusRequestDTO request,Long loggedInDealerId) throws AccessDeniedException {
 
         Vehicle vehicle = vehicleRepository.findById(id).orElseThrow(() ->
                         new ResourceNotFoundException(
                                 "Vehicle not found"));
+
+        if (!vehicle.getDealer().getId()
+                .equals(loggedInDealerId)) {
+
+            throw new AccessDeniedException(
+                    "You are not authorized to update this vehicle");
+        }
 
         if (!request.getStatus().equalsIgnoreCase("ACTIVE")
                 && !request.getStatus().equalsIgnoreCase("INACTIVE") && !request.getStatus().equalsIgnoreCase("FEATURED")) {
@@ -448,9 +460,17 @@ public class VehicleServiceImpl implements VehicleService {
 
     @Override
     @Transactional
-    public void deleteVehicle(Long id) {
+    public void deleteVehicle(Long id,Long loggedInDealerId) throws AccessDeniedException {
+
         Vehicle vehicle = vehicleRepository.findById(id).orElseThrow(() ->
                 new ResourceNotFoundException("Vehicle not found with id: " + id));
+
+        if (!vehicle.getDealer().getId()
+                .equals(loggedInDealerId)) {
+
+            throw new AccessDeniedException(
+                    "You are not authorized to delete this vehicle");
+        }
 
         leadRepository.deleteLeadsByVehicleId(vehicle.getId());
 
@@ -526,12 +546,20 @@ public class VehicleServiceImpl implements VehicleService {
     }
 
     @Override
-    public VehicleResponseDTO getVehicleById(Long vehicleId) {
+    public VehicleResponseDTO getVehicleById(Long vehicleId,Long loggedInDealerId) throws AccessDeniedException {
 
         Vehicle vehicle = vehicleRepository.findById(vehicleId)
                 .orElseThrow(() ->
                         new ResourceNotFoundException(
                                 "Vehicle not found with id : " + vehicleId));
+
+
+        if (!vehicle.getDealer().getId()
+                .equals(loggedInDealerId)) {
+
+            throw new AccessDeniedException(
+                    "You are not authorized to delete this vehicle");
+        }
 
 
         return VehicleResponseDTO.builder()
